Fix 1.9.3 version and clean some code.
Remove files that dont need to be in this (perhaps an other magical git add . ?)
add a gitignore file
add utf8 informations.

Add necessary voice that use by the Robot and instruction to be abble to
setup project

This commit was done to sync with @ben2001 to be abble to work together
to put this on a RasberryPI tomorrow morning.